+++ pkg/BonusCalc/internal/models/accrual.go
@@ -1,7 +1,7 @@
 package models
 
 type Accrual struct {
-	Order   string `json:"order"`   // Уникальный номер заказа
-	Status  string `json:"status"`  // Статус обработки
-	Accrual int    `json:"accrual"` // начисленные бонусы
+	Order   string  `json:"order"`   // Уникальный номер заказа
+	Status  string  `json:"status"`  // Статус обработки
+	Accrual float64 `json:"accrual"` // начисленные бонусы
 }
